requirements.txt
websocket-client
pyyaml
requests
